+++ CHANGELOG.md
@@ -14,12 +14,13 @@
 - Added a browser link to Build name [#1110](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1110)
 - Added a browser link to Deployment code [#1111](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1111)
 - Filter out Environments by Status [#1112](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1112)
+- Use `CCv1` api to get subscription permissions [#1113](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1113)
 
 ### `Project Import` enhancements
 - Loading backoffice sources provided by SAP on project import [#1096](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1096)
 
 ### Other
-- Migrated to `IntelliJ Platform Gradle Plugin 2.x` [#1099](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1099)
+- Migrated to `IntelliJ Platform Gradle Plugin 2.beta1` [#1099](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1099)
 
 ## [2024.1.1]